camera timer fix
git-svn-id: svn://84.194.112.164/bengine@334 2ee6282c-0497-4b59-8b7c-bcac2105c061

+++ src/utils/events.h
@@ -23,9 +23,9 @@ struct event
 	bool			active;
 	long unsigned int 	bindkey;
 	unsigned int		responsetime;
-	float			fresponsetime;		// floating responsetime
-	unsigned int		minfresponsetime;	// minimum floating responsetime
-	float			fresponseinterval;	// floating degredation interval
+	int			fresponsetime;		// responsetime
+	unsigned int		minfresponsetime;	// minimum responsetime
+	unsigned int		fresponseinterval;	// degredation interval
 	unsigned int		elapsed;
 	
 	bool			timerisshared;
